# Api/openapi.yaml
openapi: 3.0.0
info:
  title: Toolbox API
  version: 1.0.0
paths:
  /signals:
    post:
      summary: Create a new Signal
      tags:
        - Signals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignalCreateRequest"
      responses:
        "201":
          description: Signal successfully created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
    get:
      summary: Get all Signals
      tags:
        - Signals
      responses:
        "200":
          description: A list of Signals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Signal"
  /signals/{id}:
    get:
      summary: Get Signal by ID
      tags:
        - Signals
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Signal found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
        "404":
          description: Signal not found
...
components:
  schemas:
    Signal:
      type: object
      required:
        - id
        - name
        - input
        - output
        - unit
        - dataType
        - scope
        - createdAt
        - createdBy
      properties:
        id:
          type: string
        name:
          type: string
        input:
          type: boolean
          description: "Signal can be used as input for SignalProcessor"
        output:
          type: boolean
          description: "Signal can be produced as output by SignalProcessor"
        unit:
          type: string
        dataType:
          type: string
          enum: [Numeric, String]
          description: |
            Numeric → Decimal  
            String → Text value
        scope:
          type: string
          enum: [Global, Edge, Tenant]
          description: |
            - Global  
            - Edge  
            - Tenant (example: AEW)
        edgeInstance:
          type: string
          nullable: true
          description: "Optional: further limits Scope to a specific Edge Instance"
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
    SignalCreateRequest:
      type: object
      required:
        - id
        - name
        - input
        - output
        - unit
        - dataType
        - scope
      properties:
        id:
          type: string
        name:
          type: string
        input:
          type: boolean
        output:
          type: boolean
        unit:
          type: string
        dataType:
          type: string
          enum: [Numeric, String]
        scope:
          type: string
          enum: [Global, Edge, Tenant]
        edgeInstance:
          type: string
          nullable: true
...